Add Student GS File (#9)
## Summary of Changes
In this PR, I have included a new google script file that adds students to the organization as soon as they submit their usernames on the initial google form. 

## Motivation
Each semester, students need write access to the git-java-setup repository to complete Project 0. The facilitators usually do this manually but, by automating this process, we can grant students immediate access to the organization and student team. 

Closes https://github.com/sagars729/CMSC389T-Toolbox/issues/8

## Testing
This script was tested in the Fall 22 semester to add students on form submit. It successfully added new students to the organization and their team.