Add Pull Requests and Issues Project Autograder (#17)
## Summary of Changes
In this PR, I have included a new autograder for the `Pull Requests and
Issues` project. This includes updates to include a GitHub token and
Student roster to ensure that all tests run and validate successfully.

## Motivation
The Pull Requests and Issues project allows students to get early
feedback on whether they are correctly creating pull requests and issues
on GitHub. We can automate this process to give students quicker
feedback on their assignment.

Closes https://github.com/sagars729/CMSC389T-Toolbox/issues/14 

## Testing
This autograder was tested with sample submissions and deployed for
students to use in Fall 2022. Students were able to successfully get
their assignments accurately graded by using this system.