Add PacMan CI Autograder (#23)
## Summary of Changes

In this pull request, I have included a new PacManCI Autograder for the
second group project in CMSC388T

## Motivation

Students will be able to use this autograder to get early feedback on
their projects. It will also reduce the workload of graders.

## Testing
Changes were deployed and tested in the Fall 2022 semester.