Add Personal Website Autograder (#24)
## Summary of Changes
In this PR, I have included changes to create the Personal Website
Autograder. This autograder will test that student websites and actions
are working correctly on GitHub Pages, Terpconnect, and Docker through
Gradescope.

## Motivation
By adding an autograder for this project, students will be able to get
early feedback on their submission. This will also reduce the grading
workload for TAs.

## Testing
All changes were tested with students in the Fall 2022 semester.